exercices/tp_3/requirements.txt
# Streamlit
streamlit

# API
fastapi
pydantic
uvicorn
pydantic
requests
python-dotenv==1.0.1
langchain_google_genai
langchain

# LangChain pour la gestion des LLM (Large Language Models)
langchain
openai
google-cloud

